apps/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1
hiredis==2.3.2

# LangChain & LangGraph
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.20
langgraph==0.0.20
langsmith==0.0.83

# Document Parsing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.4
markdown==3.5.1

# Export
weasyprint==60.2
pypandoc==1.12

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bleach==6.1.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Code Quality
ruff==0.1.14
black==23.12.1
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-magic==0.4.27
aiofiles==23.2.1
tenacity==8.2.3
